+++ client/app/modules/workshops/workshops.module.js
@@ -5,12 +5,11 @@ import UibAlertDirective from 'angular-ui-bootstrap/src/alert';
 import HtmlIdFilter from '../../filters/htmlid/htmlid.filter';
 import TweetComponent from '../../components/tweet/tweet.component';
 import JsonLdComponent from '../../components/jsonld/jsonld.component';
-import UpcomingFilter from '../../filters/upcoming/upcoming.filter';
 import { WorkshopsComponent } from './workshops.component';
 import { WorkshopService } from '../../services/workshop.service';
 import { NewsletterService } from '../../services/newsletter.service';
 
-export default angular.module('shyApp.workshops', [ngRoute, UibAlertDirective, HtmlIdFilter, TweetComponent, JsonLdComponent, UpcomingFilter])
+export default angular.module('shyApp.workshops', [ngRoute, UibAlertDirective, HtmlIdFilter, TweetComponent, JsonLdComponent])
   .config(routes)
   .service('WorkshopService', WorkshopService)
   .service('NewsletterService', NewsletterService)